# openapi.yaml
openapi: 3.0.0
info:
  title: cineplus API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3009/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /pelicula:
    get:
      tags:
        - Pelicula
      summary: Listar todos los Peliculas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Peliculas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Pelicula
      summary: Crear Pelicula
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Pelicula creado
  /pelicula/{id}:
    get:
      tags:
        - Pelicula
      summary: Obtener Pelicula por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pelicula encontrado
        '404':
          description: Pelicula no encontrado
    patch:
      tags:
        - Pelicula
      summary: Actualizar Pelicula
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Pelicula actualizado
    delete:
      tags:
        - Pelicula
      summary: Eliminar Pelicula
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pelicula eliminado
  /pelicula/{id}/upload-image:
    post:
      tags:
        - Pelicula
      summary: Subir imagen a Pelicula
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /licencia:
    get:
      tags:
        - Licencia
      summary: Listar todos los Licencias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Licencias
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Licencia
      summary: Crear Licencia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Licencia creado
  /licencia/{id}:
    get:
      tags:
        - Licencia
      summary: Obtener Licencia por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Licencia encontrado
        '404':
          description: Licencia no encontrado
    patch:
      tags:
        - Licencia
      summary: Actualizar Licencia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Licencia actualizado
    delete:
      tags:
        - Licencia
      summary: Eliminar Licencia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Licencia eliminado
  /licencia/{id}/upload-image:
    post:
      tags:
        - Licencia
      summary: Subir imagen a Licencia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /entrada:
    get:
      tags:
        - Entrada
      summary: Listar todos los Entradas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Entradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Entrada
      summary: Crear Entrada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Entrada creado
  /entrada/{id}:
    get:
      tags:
        - Entrada
      summary: Obtener Entrada por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrada encontrado
        '404':
          description: Entrada no encontrado
    patch:
      tags:
        - Entrada
      summary: Actualizar Entrada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Entrada actualizado
    delete:
      tags:
        - Entrada
      summary: Eliminar Entrada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrada eliminado
  /entrada/{id}/upload-image:
    post:
      tags:
        - Entrada
      summary: Subir imagen a Entrada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...